make app require keys.php
All the keys are now hidden in the required keys.php file which is
gitignored. This means that everyone will have to create a keys.php file
manually in their own environment. Read the updated Readme.md for
details.

For deploy we will have to manually create a keys.php in the production
environment too, of course (but not in Git!).